pyproject.toml
[project]
name = "asimov-academy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.20",
    "ddgs>=9.10.0",
    "fastapi>=0.127.0",
    "markdown>=3.10",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
